+++ packages/plugins/type-distributor/CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
-### Added
+### Added 
+- Dialog to create S-Ieds and Aps
 - S-IED AccessPoints view
 
 ## [0.1.3] - 2026-01-19